Reorganized and created new vues
Organized vues based on before user logs in and after user logs in. Also created a forgot password and new user vue.
