+++ src/renderer/components/back/Profile-Node.vue
@@ -50,7 +50,7 @@
 
 <script type="text/javascript">
 import firebase from 'firebase'
-import Router from '../router'
+import Router from '../../router'
 
 export default {
 	data() {
