+++ src/renderer/components/front/Forgot.vue
@@ -1,0 +1,11 @@
+<template>
+	
+</template>
+
+<style type="text/css" lang="scss">
+	
+</style>
+
+<script type="text/javascript">
+	
+</script>
