+++ src/renderer/components/front/initialize.js
