+++ src/renderer/router/index.js
@@ -2,8 +2,10 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import firebase from 'firebase'
 
-import Dashboard from '@/components/Dashboard'
-import Login from '@/components/Login'
+import Dashboard from '@/components/back/Dashboard'
+import Login from '@/components/front/Login'
+import Forgot from '@/components/front/Forgot'
+import NewUser from '@/components/front/NewUser'
 
 Vue.use(Router)
 
@@ -21,6 +23,16 @@ var router = new Router({
       path: '/login',
       name: 'login',
       component: Login
+    },
+    {
+      path: '/forgot',
+      name: 'forgot-pass',
+      component: Forgot
+    },
+    {
+      path: '/newuser',
+      name: 'new-user',
+      component: NewUser
     },
     {
       path: '/dashboard',
